% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/script.R
\name{add_comment}
\alias{add_comment}
\title{Comment on an existing issue}
\usage{
add_comment(issue, comment, jira_url = getOption("jira_url"),
  jira_user = getOption("jira_user"),
  jira_password = getOption("jira_password"),
  verbose = getOption("jira_verbose"))
}
\arguments{
\item{issue}{An existing issue id or key.}

\item{comment}{Charactervector, list or data.frame. See examples.}

\item{jira_url}{base url to JIRA. Defaults to 'jira/'}

\item{jira_user}{username for authentication}

\item{jira_password}{password for authentication}

\item{verbose}{FALSE}
}
\value{
POST results
}
\description{
Comment on an existing issue
}

