% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{df_to_jira_table}
\alias{df_to_jira_table}
\title{Convert data.frame to JIRA markup table}
\usage{
df_to_jira_table(x)
}
\arguments{
\item{x}{a data.frame}
}
\value{
string
}
\description{
Convert data.frame to JIRA markup table
}
\examples{
cat( df_to_jira_table(head(iris)))
}

