% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/comment_functions.R
\name{get_comments}
\alias{get_comments}
\title{Get comments on an issue}
\usage{
get_comments(issue, jira_url = getOption("jira_url"),
  jira_user = getOption("jira_user"),
  jira_password = getOption("jira_password"),
  verbose = getOption("jira_verbose"))
}
\arguments{
\item{issue}{An existing issue id or key.}

\item{jira_url}{base url to JIRA. Defaults to 'jira/'}

\item{jira_user}{username for authentication}

\item{jira_password}{password for authentication}

\item{verbose}{FALSE}
}
\value{
POST results
}
\description{
Get comments on an issue
}

