% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_issues}
\alias{get_issues}
\title{Get all issues matching the parameters}
\usage{
get_issues(con, user = NULL, max_results = 50, start_at = 0,
  content = NULL, project_key = NULL)
}
\arguments{
\item{con}{the RjiraConnection object}

\item{user}{user to search for}

\item{max_results}{maximal number of results to be returned by Jira}
}
\value{
deserialised JSON issues data frame
}
\description{
Get all issues matching the parameters
}
\examples{

\dontrun{

con <- jira(scheme = "https", host = 'localhost')
get_issues(con,user = "admin")

issues <- get_issues(project = "ADM", issue = "ADM-6")

## Get list of issues assigned to user
res <- get_issues(user = "admin" , project_key = "ACME")

## Issues of admin in all projects
res <- get_issues(user = "admin")

## All issues in project ACME
res <- get_issues(project_key = "ACME")

}


}
\seealso{
\code{\link{search_url}}
}
