% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/script.R
\name{get_user_issues}
\alias{get_user_issues}
\title{Get issues assigned to a user}
\usage{
get_user_issues(user, jira_url = getOption("jira_url"),
  jira_user = getOption("jira_user"),
  jira_password = getOption("jira_password"))
}
\arguments{
\item{user}{user to search for}

\item{jira_url}{base JIRA url}

\item{jira_user}{username for authentication}

\item{jira_password}{password for authentication}
}
\value{
character vector of issues
}
\description{
Get issues assigned to a user
}
\examples{
get_user_issues(user = "billy")
}
\seealso{
\code{\link{search_url}}
}

