% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/script.R
\name{get_watchers}
\alias{get_watchers}
\title{Get a list of watchers on an existing issue}
\usage{
get_watchers(issue, jira_url = getOption("jira_url"),
  jira_user = getOption("jira_user"),
  jira_password = getOption("jira_password"),
  verbose = getOption("jira_verbose"))
}
\arguments{
\item{jira_url}{base url to JIRA. Defaults to 'jira/'}

\item{jira_user}{username for authentication}

\item{jira_password}{password for authentication}

\item{verbose}{FALSE}
}
\value{
DELETE results
}
\description{
Get a list of watchers on an existing issue
}
\seealso{
\code{\link{add_watcher}}

\code{\link{remove_watcher}}
}

