% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/script.R
\name{issue_url}
\alias{issue_url}
\title{Get jira issue url}
\usage{
issue_url(jira_url = getOption("jira_url"))
}
\arguments{
\item{jira_url}{base url to jira. Defaults to 'jira/'}
}
\value{
string
}
\description{
Get jira issue url
}
\examples{
issue_url()
}

