% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{jira}
\alias{jira}
\title{Connect to a jira instance}
\usage{
jira(scheme = c("http", "https"), host = "localhost", port = 8080,
  user = "user", pass = "pass", project = "/", verbose = FALSE,
  curl_options = NULL)
}
\arguments{
\item{jira_url}{base url to jira. Defaults to 'jira/'}
}
\description{
Connect to a jira instance
}
\examples{
jira()
}
