% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jr_get_meta.R
\name{jr_get_createmeta}
\alias{jr_get_createmeta}
\title{Get create issue metadata}
\usage{
jr_get_createmeta()
}
\value{
table with metadata
}
\description{
Get create issue metadata
}
\details{
See \href{https://developer.atlassian.com/cloud/jira/platform/rest/v3/api-group-issues/#api-rest-api-3-issue-createmeta-get}{documents}
}
