% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jr_get_projects.R
\name{jr_get_projects}
\alias{jr_get_projects}
\title{Get all projects}
\usage{
jr_get_projects()
}
\value{
tibble with projects info
}
\description{
Get all projects
}
\details{
See \href{https://developer.atlassian.com/cloud/jira/platform/rest/v3/api-group-projects/#api-rest-api-3-project-get}{documentation}
}
