% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jr_get_updated_worklogs.R
\name{jr_get_updated_worklogs}
\alias{jr_get_updated_worklogs}
\title{Get IDs of updated worklogs}
\usage{
jr_get_updated_worklogs(since, expand = "properties")
}
\arguments{
\item{since}{The date and time, after which updated worklogs are returned.}

\item{expand}{Use expand to include additional information about worklogs in the response. This parameter accepts properties that returns the properties of each worklog.}
}
\value{
tibble with IDs and update timestamps for worklogs
}
\description{
Returns a list of IDs and update timestamps for worklogs updated after a date and time.
For more information see next \href{https://developer.atlassian.com/cloud/jira/platform/rest/v3/api-group-issue-worklogs/#api-rest-api-3-worklog-updated-get}{link}.
}
\examples{
\dontrun{
# get worklogs ids updated yesterday
updated_wl <- jr_get_updated_worklogs(Sys.Date()-1)
}
}
