% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jr_get_worklogs.R
\name{jr_get_worklogs}
\alias{jr_get_worklogs}
\title{Get worklogs}
\usage{
jr_get_worklogs(ids, expand = "properties")
}
\arguments{
\item{ids}{A vector of worklog IDs.}

\item{expand}{Use expand to include additional information about worklogs in the response. This parameter accepts properties that returns the properties of each worklog.}
}
\value{
tibble with worklog details
}
\description{
Returns worklog details for a list of worklog IDs.
}
\examples{
\dontrun{
# get worklogs ids updated yesterday
updated_wl <- jr_get_updated_worklogs(Sys.Date()-1)
# get worklogs details
worklogs <- jr_get_worklogs(updated_wl$worklog_id)

}
}
