% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{project_url}
\alias{project_url}
\title{Get jira project url}
\usage{
project_url(con)
}
\arguments{
\item{jira_url}{base url to jira. Defaults to 'jira/'}
}
\value{
string
}
\description{
Get jira project url
}
\examples{
issue_url()
}
