% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/watcher_functions.R
\name{remove_watcher}
\alias{remove_watcher}
\title{Remove a watched to an existing issue}
\usage{
remove_watcher(issue, user, jira_url = getOption("jira_url"),
  jira_user = getOption("jira_user"),
  jira_password = getOption("jira_password"),
  verbose = getOption("jira_verbose"))
}
\arguments{
\item{jira_url}{base url to JIRA. Defaults to 'jira/'}

\item{jira_user}{username for authentication}

\item{jira_password}{password for authentication}

\item{verbose}{FALSE}
}
\value{
DELETE results
}
\description{
Remove a watched to an existing issue
}
\seealso{
\code{\link{add_watcher}}

\code{\link{get_watchers}}
}

