% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjira-package.R, R/rjira_package.R
\docType{package}
\name{rjira-package}
\alias{rjira-package}
\title{rjira: Work with Jira API}
\description{
Load data from Jira API \url{https://developer.atlassian.com/cloud/jira/platform/rest/v3/intro/#about}.

Load data from Jira API \url{https://developer.atlassian.com/cloud/jira/platform/rest/v3/intro/#about}.
}
\author{
\strong{Maintainer}: Alexey Seleznev \email{selesnow@gmail.com} (\href{https://orcid.org/0000-0003-0410-7385}{ORCID})

}
\keyword{internal}
