% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rjira.getJQL.project.r
\name{Rjira.getJQL.project}
\alias{Rjira.getJQL.project}
\title{Get JQL query results for specific project.}
\usage{
Rjira.getJQL.project(
  startAt = 0,
  project,
  maxResults = 100,
  responseFormat = "parsed"
)
}
\arguments{
\item{startAt}{- numeric. The index number of results start page to fetch. By default is set to '0', to point at the very 1'st issue.}

\item{project}{- the name of the project issues should be fetched for.}

\item{maxResults}{- numeric. The number of issues that should be fetched in a single request. By default is set to '100' which is the maximal number of issues allowed by JIRA API.}

\item{responseFormat}{- format of the API response: parsed / raw / text. Default is "parsed".}
}
\description{
Get JQL query results for specific project.
}
