% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rjira.getJQL.project.totalIssues.r
\name{Rjira.getJQL.project.totalIssues}
\alias{Rjira.getJQL.project.totalIssues}
\title{Get the length of the whole issues list in given project.}
\usage{
Rjira.getJQL.project.totalIssues(project, responseFormat = "parsed")
}
\arguments{
\item{project}{- the name of the project the list of issues should be fetched for.}

\item{responseFormat}{- format of the API response: parsed / raw / text}
}
\value{
The number of total issues in project.
}
\description{
Get the length of the whole issues list in given project.
}
