% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rjira.getJQL.project.totalIssues.r
\name{Rjira.getJQL.project.totalIssues.update}
\alias{Rjira.getJQL.project.totalIssues.update}
\title{Get the length of the issues list in given project updated within the given period of time.}
\usage{
Rjira.getJQL.project.totalIssues.update(
  project,
  update,
  responseFormat = "parsed"
)
}
\arguments{
\item{project}{- The name of the project the list of issues should be fetched for.}

\item{update}{- Code of a time period to determine the update window. 'Xh' - X hours, 'Xd' - X days, 'Xw' - X weeks, 'Xm' - X months. Ie. '5d' will only fetch issues updated within last 5 days.}
}
\value{
- Number of total issues within a project that were updated within the given period of time.
}
\description{
Get the length of the issues list in given project updated within the given period of time.
}
