% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rjira.getJQL.project.r
\name{Rjira.getJQL.project.update}
\alias{Rjira.getJQL.project.update}
\title{Get JQL query results for specific project, updated within given period of time.}
\usage{
Rjira.getJQL.project.update(
  startAt,
  project,
  maxResults = 100,
  update,
  responseFormat = "parsed"
)
}
\arguments{
\item{startAt}{- numeric. The index number of results start page to fetch. By default is set to '0', to point at the very 1'st issue.}

\item{project}{- the name of the project issues should be fetched for.}

\item{maxResults}{- numeric. The number of issues that should be fetched in a single request. By default is set to '100' which is the maximal number of issues allowed by JIRA API.}

\item{update}{- code of a time period to determine the update window. 'Xh' - X hours, 'Xd' - X days, 'Xw' - X weeks, 'Xm' - X months. Ie. '5d' will only fetch issues updated within last 5 days.}

\item{responseFormat}{- format of the API response: parsed / raw / text.}
}
\description{
Get JQL query results for specific project, updated within given period of time.
}
