% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rjira.getJQL.rawQuery.r
\name{Rjira.getJQL.rawQuery}
\alias{Rjira.getJQL.rawQuery}
\title{Get results for JQL query}
\usage{
Rjira.getJQL.rawQuery(query, responseFormat = "parsed", tryAttempts = 5)
}
\arguments{
\item{query}{- JQL query. For full reference please refer to Atlassian Cloud: https://support.atlassian.com/jira-software-cloud/docs/what-is-advanced-searching-in-jira-cloud/}

\item{responseFormat}{- format of the API response: parsed / raw / text}

\item{tryAttempts}{- how many times should JIRA API be querioed if response will not be given immediately. By default 5 attempts are performed}
}
\value{
- Raw JSON response from JIRA API. For further processing it should be parse by Rjira.parseissues function.
}
\description{
This is the basic, low level request function used by other high level functions in this package. You can use it to get raw results for your query, or use othe high level functions to get specific results for either project or category of issues.
}
