% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rjira.jobs.getIssues.r
\name{Rjira.jobs.getIssues.all}
\alias{Rjira.jobs.getIssues.all}
\title{High level function to get all issues from your JIRA database for a specific project.}
\usage{
Rjira.jobs.getIssues.all(project, save = TRUE)
}
\arguments{
\item{project}{- Your JIRA project name your wish to fetch issues from.}

\item{save}{- TRUE or FALSE whether to save your output to database. Default - TRUE.}
}
\value{
- List of parsed issues from your JIRA project.
}
\description{
Before proceeding please run Rjira.utils.storeJiraCredentials and Rjira.utils.storeOdbcCredentials if you want to save your output to database.
}
