% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rjira.jobs.getIssues.r
\name{Rjira.jobs.getIssues.updated}
\alias{Rjira.jobs.getIssues.updated}
\title{High level function to get all issues from your JIRA database for a specific project, updated within a specific timeframe.}
\usage{
Rjira.jobs.getIssues.updated(project, update, save = TRUE)
}
\arguments{
\item{project}{- Your JIRA project name your wish to fetch issues from.}

\item{update}{- Code of a time period to determine the update window. 'Xh' - X hours, 'Xd' - X days, 'Xw' - X weeks, 'Xm' - X months. Ie. '5d' will only fetch issues updated within last 5 days.}

\item{save}{- TRUE or FALSE whether to save your output to database. Default - TRUE.}
}
\value{
- List of parsed issues from your JIRA project.
}
\description{
Before proceeding please run Rjira.utils.storeJiraCredentials and Rjira.utils.storeOdbcCredentials if you want to save your output to database.
}
