% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rjira.parseIssues.customFieldsDictionary.r
\name{Rjira.parseIssues.customFieldsDictionary}
\alias{Rjira.parseIssues.customFieldsDictionary}
\title{Build custom fields dictionary from JIRA API response.}
\usage{
Rjira.parseIssues.customFieldsDictionary(response)
}
\arguments{
\item{response}{- The parsed response returned from one of Rjira.getJQL.rawQuery or Rjira.getJQL.project functions.}
}
\value{
A tibble containing dictionary for custom fields mentioned in the JIRA API response.
}
\description{
Build custom fields dictionary from JIRA API response.
}
