% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rjira.parseIssues.r
\name{Rjira.parseIssues}
\alias{Rjira.parseIssues}
\title{Parse issues on a list returned from Rjira.getJQL.rawQuery or Rjira.getJQL.project.}
\usage{
Rjira.parseIssues(response, namesDict, save = TRUE)
}
\arguments{
\item{response}{- Parsed response from Rjira.getJQL.project or Rjira.getJQL.project.}

\item{namesDict}{- The dictionary of custom fields as returned from Rjira.parseIssues.customFieldsDictionary.}

\item{save}{- TRUE or FALSE on whether to save the results in the database.}
}
\value{
- the list of parsed issues with custom fields with friendly names.
}
\description{
Parse issues on a list returned from Rjira.getJQL.rawQuery or Rjira.getJQL.project.
}
