% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rjira.utils.storeJiraCredentials.r
\name{Rjira.utils.storeJiraCredentials}
\alias{Rjira.utils.storeJiraCredentials}
\title{Store Rjira_credentials object to local Environment.}
\usage{
Rjira.utils.storeJiraCredentials()
}
\description{
The object should contain your JIRA domain, login and pass which will be used in Rjira functions globally.
}
