% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rjira.utils.storeOdbcCredentials.r
\name{Rjira.utils.storeOdbcCredentials.win}
\alias{Rjira.utils.storeOdbcCredentials.win}
\title{Store Rjira_odbcCredentials object to local Environment.}
\usage{
Rjira.utils.storeOdbcCredentials.win()
}
\description{
The object should contain your odbc credentials if you want to save Rjira pipeline results to SQL database. These information include on Windows - login, pass and DNS name. These will be used in Rjira functions globally.
}
