% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/script.R
\name{search_url}
\alias{search_url}
\title{Get jira search url}
\usage{
search_url(jira_url = getOption("jira_url"))
}
\arguments{
\item{jira_url}{base url to jira. Defaults to 'jira/'}
}
\value{
string
}
\description{
Get jira search url
}
\examples{
search_url()
}

