% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/comment_functions.R
\name{add_comment}
\alias{add_comment}
\title{Comment on an existing issue}
\usage{
add_comment(issue, comment, jira_url = getOption("jira_url"),
  jira_user = getOption("jira_user"),
  jira_password = getOption("jira_password"),
  verbose = getOption("jira_verbose"))
}
\arguments{
\item{issue}{An existing issue id or key.}

\item{comment}{Character vector, list or data.frame. See examples.}

\item{jira_url}{base url to JIRA. Defaults to 'jira/'}

\item{jira_user}{username for authentication}

\item{jira_password}{password for authentication}

\item{verbose}{FALSE}
}
\value{
POST results
}
\description{
Comment on an existing issue
}
\examples{
\dontrun{

## set options
 options("jira_user" = "admin") ## my username is admin
 options("jira_password" = "xxxx")
 options("jira_url" = "jira/")
 options("jira_project" = "BAS")

## write comment to issue
add_comment(issue = "BAS-1", comment = "A comment")

## comment with formatting
my_comment <- "
 Bullet points example
 * point 1
 ** point 1.1
 * point 2
"
add_comment(issue = "BAS-1", comment = my_comment) ## formatting works


## write a data.frame as a table
my_results_comment <- list("The results of this analysis:", head(iris, 4))
add_comment(issue = "BAS-1", comment = my_results_comment) ## formatting works

}
}

