% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_issues}
\alias{get_issues}
\title{Get issues assigned to a user}
\usage{
get_issues(user = NULL, project_key = getOption("jira_project"),
  jira_url = getOption("jira_url"), jira_user = getOption("jira_user"),
  jira_password = getOption("jira_password"),
  verbose = getOption("jira_verbose"))
}
\arguments{
\item{user}{user to search for}

\item{jira_url}{base JIRA url}

\item{jira_user}{username for authentication}

\item{jira_password}{password for authentication}
}
\value{
character vector of issues
}
\description{
Get issues assigned to a user
}
\examples{
get_issues(user = "admin")
}
\seealso{
\code{\link{search_url}}
}

