% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_issues}
\alias{get_issues}
\title{Get issues assigned to a user}
\usage{
get_issues(user = NULL, project_key = getOption("jira_project"),
  jira_url = getOption("jira_url"), jira_user = getOption("jira_user"),
  jira_password = getOption("jira_password"),
  verbose = getOption("jira_verbose"))
}
\arguments{
\item{user}{user to search for}

\item{jira_url}{base JIRA url}

\item{jira_user}{username for authentication}

\item{jira_password}{password for authentication}
}
\value{
character vector of issues
}
\description{
Get issues assigned to a user
}
\examples{
\dontrun{

get_issues(user = "admin")

issues <- get_issues(project = "ADM", issue = "ADM-6", verbose = T)

## Get list of issues assigned to user
res <- get_issues(user = "admin" , project_key = getOption("jira_project"))
sapply(res, "[[", "id")
sapply(res, "[[", "key")

## All projects
res <- get_issues(user = "admin" , project_key = NULL)

}
}
\seealso{
\code{\link{search_url}}
}

