% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_projects}
\alias{get_projects}
\title{Get all projects}
\usage{
get_projects(jira_url = getOption("jira_url"),
  jira_user = getOption("jira_user"),
  jira_password = getOption("jira_password"),
  verbose = getOption("jira_verbose"))
}
\arguments{
\item{jira_url}{base JIRA url}

\item{jira_user}{username for authentication}

\item{jira_password}{password for authentication}

\item{verbose}{FALSE}
}
\value{
TODO
}
\description{
Get all projects
}
\examples{
get_projects()
}
\seealso{
\code{\link{search_url}}
}
