% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jr_issue_search.R
\name{jr_issue_search}
\alias{jr_issue_search}
\title{Search for issues using JQL}
\usage{
jr_issue_search(fields = "*all", expand = c("schema", "names"), jql = "")
}
\arguments{
\item{fields}{A list of fields to return for each issue, use it to retrieve a subset of fields. This parameter accepts a comma-separated list.}

\item{expand}{Use expand to include additional information about issues in the response. Note that, unlike the majority of instances where expand is specified, expand is defined as a list of values. The expand options are:}

\item{jql}{A JQL expression. See \href{https://confluence.atlassian.com/x/egORLQ}{documentation}}
}
\value{
tibble with search result
}
\description{
Search for issues using JQL
}
\details{
You can see details \href{https://developer.atlassian.com/cloud/jira/platform/rest/v3/api-group-issue-search/#api-rest-api-3-search-post}{this link}
}
